chore: move sidebar inside main-layout folder
