+++ frontend/src/app/main-layout/sidebar/sidebar.component.html
