+++ frontend/src/app/main-layout/sidebar/sidebar.component.spec.ts
